create http server module
