+++ src/server/timelines/game-data.ts
@@ -15,11 +15,12 @@ import { RoomName } from "@server/game-data/rooms";
 import { getStagePlayMusic, StageScript } from "@server/game-data/stage-plays";
 import { ORIGINAL_STAMPBOOK, Stampbook, StampCategory, StampRoom, STAMP_ROOMS } from "@server/game-data/stamps";
 import { FURNITURE } from "@server/game-logic/furniture";
-import { ItemTable } from "@server/game-logic/items";
+import { ITEMS, ItemTable } from "@server/game-logic/items";
 import { WaddleRoomInfo } from "@server/game-logic/waddles";
 import { isGreater, Version } from "@server/routes/versions";
 import { SettingsManager } from "@server/settings";
 import { CatalogItems, CPUpdateE, CrumbIndicator, GameUpdate, HuntCrumbs, IglooList, ListSongPatch, PartyOp, WorldStamp } from "@server/updates";
+import { getUpdates } from "@server/updates/updates";
 import path from "path";
 
 const SCAVENGER_ICON_PATH = 'scavenger_hunt/scavenger_hunt_icon.swf';
@@ -165,7 +166,14 @@ export class GameData {
 
   private date: string;
 
-  constructor(private updates: GameUpdate[], private items: ItemTable, settings: SettingsManager) {
+  private updates: GameUpdate[]
+
+  private items: ItemTable;
+
+  constructor(settings: SettingsManager) {
+    this.updates = getUpdates();
+    // todo: remove global state
+    this.items = ITEMS;
     this.date = settings.settings.version;
     this.update(settings.settings.version);
     settings.addListener(() => {
